lib/gstreamer: add common transcoder base
Add a transcoder base class for gstreamer to share
common code between msdk and vaapi elements.
